Major refactoring, simplify java code

+++ app/isoparser/src/main/java/com/coremedia/iso/boxes/TrackReferenceTypeBox.java
@@ -42,7 +42,7 @@ public class TrackReferenceTypeBox extends AbstractBox {
 
     @Override
     public void _parseDetails(ByteBuffer content) {
-        int count = (int) (content.remaining() / 4);
+        int count = content.remaining() / 4;
         trackIds = new long[count];
         for (int i = 0; i < count; i++) {
             trackIds[i] = IsoTypeReader.readUInt32(content);
